backend/requirements.txt
djangorestframework
Django>=4.2
python-dotenv
PyMySQL
djangorestframework
